Custom KeyValuePair class, other work
Added custom KeyVal class to allow for mutable keys and values.

Started adding more to RollLineSimple so that it can handle the same
data as the full RollLine, but with a lighter footprint

ShotChart now populates it's arrays when a new instance is created

Small bits of work on DiceRoller, starting to change over to new system.